Rough draft of a standalone "binary"
Needs some refactoring, tests, better input handling and documentation, but is
a functional proof-of-concept. Built a nearly-equivalent index using:

$ bundle exec bin/jekyll_pages_api "" ../hub/_site \
  "18F &mdash; " '<div class="bare-content" ' > out.txt

I say "nearly-equivalent" because in the new version, URLs don't have trailing
slashes, and the "Edit this link" text appears at the end of the body. We can
update the layout to trim the latter. Also, files without titles aren't
included, such as the auth-include snippets.