Test that tags and title are in JSON metadata
After discussion IRL with Aidan, add tests for the `tags` and `title`
metadata in the JSON. This is to document that we intentionally
duplicate the tags and title.

We duplicate this because: as a human it makes sense to reduce
duplication, but as a machine I would expect to see all the frontmatter
in one place.